pyproject.toml
[project]
name = "scrsit"
version = "0.1.0"
description = "Smart(Intelligent) Customer Requirements Interpretation"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "Apache License Version 2.0"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
